% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{tr_sql_query}
\alias{tr_sql_query}
\title{Executa uma consulta somente de leitura.}
\usage{
tr_sql_query(fonte, consulta)
}
\arguments{
\item{fonte}{Valor produzido por [tr_sql_source()].}

\item{consulta}{Instrução SQL iniciada por SELECT ou WITH.}
}
\value{
Tabela data.frame com o resultado.
}
\description{
Executa uma consulta somente de leitura.
}
